# Conflicts:
#	src/engine/DrawManager.java

+++ src/engine/GameState.java
@@ -13,13 +13,15 @@ public class GameState {
 	/** Current score. */
 	private int score;
 	/** Lives currently remaining. */
-	private int livesRemaining;
+	private double livesRemaining;
 	/** Bullets shot until now. */
 	private int bulletsShot;
 	/** Ships destroyed until now. */
 	private int shipsDestroyed;
 	/** HardCore(Only One life) */
 	private boolean hardcore;
+
+
 
 	/**
 	 * Constructor.
@@ -38,7 +40,7 @@ public class GameState {
 	 *            Hardcore mode, Only one coin.
 	 */
 	public GameState(final int level, final int score,
-			final int livesRemaining, final int bulletsShot,
+			final double livesRemaining, final int bulletsShot,
 			final int shipsDestroyed, final boolean hardcore) {
 		this.level = level;
 		this.score = score;
@@ -46,6 +48,7 @@ public class GameState {
 		this.bulletsShot = bulletsShot;
 		this.shipsDestroyed = shipsDestroyed;
 		this.hardcore = hardcore;
+
 	}
 
 	/**
@@ -65,7 +68,7 @@ public class GameState {
 	/**
 	 * @return the livesRemaining
 	 */
-	public final int getLivesRemaining() {
+	public final double getLivesRemaining() {
 		return livesRemaining;
 	}
 
@@ -89,6 +92,8 @@ public class GameState {
 	public final boolean getHardCore() {
 		return this.hardcore;
 	}
+
+	//public final int getBulletsCount() { return this.BulletsCount;}
 
 	/**
 	 * Set HardCore
